+++ tests/app/models.py
@@ -1,4 +1,5 @@
 from django.db import models
+from django.conf import settings
 
 
 class City(models.Model):
@@ -21,6 +22,7 @@ class Person(models.Model):
     name = models.CharField(max_length=255)
     email = models.EmailField(blank=True)
     date = models.DateTimeField(null=True, blank=True, help_text="Select one of 'Yesterday', 'Today', 'Tomorrow'")
+    language = models.CharField(choices=settings.LANGUAGES, max_length=64, blank=True)
 
     def __str__(self):
         return self.name